// src/app/api/catalog/model/movieDetailWrapperResponseDto.ts
/**
 * CinePik Catalog API
 * The CinePik Catalog microservice.
 *
 * The version of the OpenAPI document: 0.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MovieDetailResponseDto } from './movieDetailResponseDto';
import { SimilarMovieDetailResponseDto } from './similarMovieDetailResponseDto';


export interface MovieDetailWrapperResponseDto { 
    movie: MovieDetailResponseDto;
    /**
     * Similar movies.
     */
    similarMovies: Array<SimilarMovieDetailResponseDto>;
}