// src/app/api/catalog/model/showDetailWrapperResponseDto.ts
/**
 * CinePik Catalog API
 * The CinePik Catalog microservice.
 *
 * The version of the OpenAPI document: 0.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShowDetailResponseDto } from './showDetailResponseDto';
import { ShowSeasonsResponseDto } from './showSeasonsResponseDto';


export interface ShowDetailWrapperResponseDto { 
    show: ShowDetailResponseDto;
    /**
     * Show seasons.
     */
    seasons: Array<ShowSeasonsResponseDto>;
}